// src/models/Vehicle.ts
export type Vehicle = {
    id: number;
    make: string;
    model: string;
    year: Date;
    engine: string;
    fuel: string;
    registration_number: string;
    gearbox: string;
    drivetrain: string;
    package: string;
    trim: string;
    weight: number;
    fuel_consumption_avg: number;
    odometer: number;
    img: string;
    service_interval: number;
    created_at: string
};